requirements.txt
aiohttp==3.10.5
aiohttp_socks==0.9.0
bip44==0.1.4
coincurve==18.0.0
inputimeout==1.0.4
peewee==3.17.6
python-dotenv==1.0.1
python-socks[asyncio]==2.5.1
stem==1.8.2
